CORE-8711 Move Groups theme to Collaborators theme package
