+++ de-lib/src/main/java/org/iplantc/de/theme/base/client/collaborators/GroupDisplayStrings.java
@@ -1,4 +1,4 @@
-package org.iplantc.de.theme.base.client.groups;
+package org.iplantc.de.theme.base.client.collaborators;
 
 import com.google.gwt.i18n.client.Messages;
