--- de-lib/src/main/java/org/iplantc/de/theme/base/client/groups/Groups.gwt.xml
@@ -1,7 +1,0 @@
-<module>
-    <!-- View -->
-    <replace-with class="org.iplantc.de.theme.base.client.groups.GroupViewDefaultAppearance">
-        <when-type-is class="org.iplantc.de.groups.client.GroupView.GroupViewAppearance"/>
-    </replace-with>
-
-</module>
